+++ packages/app/src/components/presentational/ChangeSkinButton/ChangeSkinForm.spec.tsx
@@ -1,17 +1,19 @@
 import * as React from "react";
 import { shallow } from "enzyme";
+import { ChangeSkinForm } from "./ChangeSkinForm";
 import * as FyghtState from "../../../state";
-import { FyghterChangeSkinModal } from ".";
 import { storeMocks } from "../../../testHelpers";
 const { myFyghter } = storeMocks;
 
-describe("FyghterChangeSkinModal", () => {
+describe("ChangeSkinForm", () => {
   test("should render the component", () => {
     jest
       .spyOn(FyghtState, "useFyghtState")
       // eslint-disable-next-line @typescript-eslint/no-explicit-any
-      .mockImplementation((): any => [{ myFyghter }, { changeMyFyghterSkin: jest.fn() }]);
+      .mockImplementation((): any => [{ myFyghter }]);
 
-    expect(shallow(<FyghterChangeSkinModal isLoading={false} />)).toMatchSnapshot();
+    expect(
+      shallow(<ChangeSkinForm visible={true} onSave={jest.fn()} onCancel={jest.fn()} errorMessage={""} />)
+    ).toMatchSnapshot();
   });
 });
