+++ packages/app/src/components/presentational/ChangeSkinButton/index.tsx
@@ -1,0 +1,1 @@
+export * from "./ChangeSkinButton";
