--- packages/app/src/components/presentational/FyghterChangeSkinModal/index.tsx
@@ -1,1 +1,0 @@
-export * from "./FyghterChangeSkinModal";
